weather-server-python/requirements.txt
pyyaml==6.0.1
rich==13.7.0
typer==0.9.0 
mcp>=1.0.0
uvicorn>=0.25.0
fastapi>=0.109.0 